requirements.txt
lxml
python-dateutil
fastapi
SQLAlchemy
flask_sqlalchemy
PyMySQL
python-dotenv
